floatCalc {rCSGOfloat}	R Documentation

Output float and condition calculation for trade-up contracts

Description

This function returns the float and condition of the skin(s) you want to trade up to. A helpful resource for finding the float boundaries is https://csgostash.com/

Usage

floatCalc(skinFloats, min, max)

Arguments

skinFloats
A numeric vector of length 10 consisting of the float values of the input skins of your trade up contract.

min
The upper boundary of the float value of the skin you want to trade up to.

max
The lower boundary of the float value of the skin you want to trade up to.

Examples

floatCalc(skinFloats = c(0.1,0.12,0.123,0.1234,0.12345,
0.98765,0.9876,0.987,0.98,0.9), min = 0, max = 1)
> Float:  0.54321 Condition: Battle-Scarred

[Package rCSGOfloat version 0.1.0 Index]