neededFloat {rCSGOfloat}	R Documentation

Average needed float calculation for trade-up contracts.

Description

Calculate the average float of input skins needed to get the output skin in the desired condition. A helpful resource for finding the float boundaries is https://csgostash.com/

Usage

neededFloat(x)

Arguments

desiredCondition
Character vector containing the desired condition of the output skin from the trade-up contract.

Format:

"FN" | "Factory New"

"MW" | "Minimal Wear"

"FT" | "Field-Tested"

"WW" | "Well-Worn"

"BS" | "Battle-Scarred"

min
The upper boundary of the float value of the skin you want to trade up to.

max
The lower boundary of the float value of the skin you want to trade up to.

Examples

neededFloat(desiredCondition = "MW",min = 0,max = 0.75)
 > Average Condition Needed:  0.192307692307692

[Package rCSGOfloat version 0.1.0 Index]